+++ flow/src/main/scala/org/alephium/flow/core/BlockHeaderPool.scala
@@ -16,6 +16,8 @@
 
 package org.alephium.flow.core
 
+import java.math.BigInteger
+
 import org.alephium.io.IOResult
 import org.alephium.protocol.BlockHash
 import org.alephium.protocol.model.BlockHeader
@@ -29,7 +31,7 @@ trait BlockHeaderPool extends BlockHashPool {
   def getBlockHeader(hash: BlockHash): IOResult[BlockHeader]
   def getBlockHeaderUnsafe(hash: BlockHash): BlockHeader
 
-  def add(header: BlockHeader, weight: BigInt): IOResult[Unit]
+  def add(header: BlockHeader, weight: BigInteger): IOResult[Unit]
 
   def getHeadersAfter(locator: BlockHash): IOResult[AVector[BlockHeader]] = {
     for {
@@ -40,7 +42,7 @@ trait BlockHeaderPool extends BlockHashPool {
 
   def getHeight(bh: BlockHeader): IOResult[Int] = getHeight(bh.hash)
 
-  def getWeight(bh: BlockHeader): IOResult[BigInt] = getWeight(bh.hash)
+  def getWeight(bh: BlockHeader): IOResult[BigInteger] = getWeight(bh.hash)
 
   def isTip(bh: BlockHeader): Boolean = isTip(bh.hash)
